+++ src/main/java/com/USWRandomChat/backend/member/api/MemberController.java
@@ -8,6 +8,7 @@ import com.USWRandomChat.backend.member.memberDTO.MemberDTO;
 import com.USWRandomChat.backend.member.memberDTO.SignInRequest;
 import com.USWRandomChat.backend.member.memberDTO.SignInResponse;
 import com.USWRandomChat.backend.member.memberDTO.SignUpRequest;
+import com.USWRandomChat.backend.member.service.FindIdService;
 import com.USWRandomChat.backend.member.service.MemberService;
 import com.USWRandomChat.backend.response.ListResponse;
 import com.USWRandomChat.backend.response.ResponseService;
@@ -37,6 +38,7 @@ public class MemberController {
     private final ResponseService responseService;
     private final EmailService emailService;
     private final JwtService jwtService;
+    private final FindIdService findIdService;
 
     //회원가입
     @PostMapping(value = "/sign-up")
@@ -147,7 +149,7 @@ public class MemberController {
     //Id 찾기 로직: 이메일 인증된 회원만
     @PostMapping(value = "/find-Id")
     public ResponseEntity<Boolean> findId(@RequestParam String email) throws MessagingException {
-        return new ResponseEntity<>(memberService.findById(email), HttpStatus.OK);
+        return new ResponseEntity<>(findIdService.findById(email), HttpStatus.OK);
     }
 
     //임의의 데이터 넣기
