+++ src/main/java/com/USWRandomChat/backend/member/service/MemberService.java
@@ -177,33 +177,4 @@ public class MemberService {
         memberRepository.deleteById(id);
     }
 
-    public boolean findById(String email) throws MessagingException {
-        Member findmember = memberRepository.findByEmail(email);
-
-        // 이메일에 해당하는 회원이 없을 때 예외 발생
-        if (findmember == null){
-            throw new AccountException(ExceptionType.BAD_CREDENTIALS);
-        }
-
-        // 이메일이 인증되지 않은 경우 예외를 던집니다.
-        if (!findmember.isEmailVerified()) {
-            throw new AccountException(EMAIL_NOT_AUTHED);
-        }
-
-        try {
-            //MimeMessage  생성
-            MimeMessage mimeMessage = javaMailSender.createMimeMessage();
-            MimeMessageHelper helper = new MimeMessageHelper(mimeMessage, false);
-            helper.setTo(findmember.getEmail() + "@suwon.ac.kr"); // 수신자 이메일 설정
-            helper.setSubject("수원대학교 SWCHAT의 아이디 찾기 위한 메일입니다."); // 이메일 제목
-            helper.setText("회원님의 아이디는 " + findmember.getAccount() + " 입니다."); // 이메일 내용 설정
-
-            javaMailSender.send(mimeMessage);
-            return true;
-        } catch (MessagingException e){
-            e.printStackTrace();
-            throw new MailException(SEND_MAIL_FAILED);
-        }
-    }
-
 }
